+++ packages/browser/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @glissade/browser
 
+## 0.44.0-pre.1
+
+### Patch Changes
+
+- @glissade/backend-canvas2d@0.44.0-pre.1
+- @glissade/backend-dom@0.44.0-pre.1
+- @glissade/core@0.44.0-pre.1
+- @glissade/element@0.44.0-pre.1
+- @glissade/player@0.44.0-pre.1
+- @glissade/scene@0.44.0-pre.1
+
 ## 0.44.0-pre.0
 
 ### Patch Changes